+++ .gitignore
@@ -36,3 +36,5 @@ free-space
 byzantine/tests/node_modules
 
 *.out
+
+node_modules/*
